issue(13): implement gallery and change dev to prod in deploy/build scripts
