+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {},
   "devDependencies": {
     "@leadgrabr/brand-badges": "0.0.2",
+    "@leadgrabr/gallery": "0.0.3",
     "@leadgrabr/lead-form": "0.0.8",
     "@leadgrabr/testimonials": "0.0.3",
     "api-bro": "^1.3.9",
@@ -73,8 +74,8 @@
   "scripts": {
     "start": "npm run start-dev",
     "dev": "better-npm-run start-dev",
-    "build-dev": "better-npm-run build-dev",
-    "deploy-dev": "npm run build-dev; aws s3 sync ./dist s3://annarbor.pictures --exclude '.DS_Store' --profile leadgrabr",
+    "build-prod": "better-npm-run build-prod",
+    "deploy-prod": "npm run build-prod; aws s3 sync ./dist s3://annarbor.pictures --exclude '.DS_Store' --profile leadgrabr",
     "lint": "eslint -c .eslintrc src"
   },
   "betterScripts": {
@@ -89,8 +90,8 @@
         "GOOGLE_MAPS_APIKEY": "AIzaSyCgDmnVgrSncTpR8vgsyEKSBUm1iXHM7Y4"
       }
     },
-    "build-dev": {
-      "command": "webpack",
+    "build-prod": {
+      "command": "webpack -p",
       "env": {
         "AUDIENCE": "570e3a7c6352af5792027c25",
         "API_ENDPOINT": "http://api.leadgrabr.com/api/v1",
