+++ src/components/Contact.js
@@ -1,7 +1,7 @@
 /* eslint-disable react/prefer-stateless-function */
 import { default as React, Component, PropTypes } from 'react'
 import { Page } from 'components'
-import { default as Gallery } from 'Gallery'
+import { default as Gallery } from '@leadgrabr/gallery'
 import { default as LeadForm } from '@leadgrabr/lead-form'
 import { connect } from 'redux-await'
 import { constants, createLead } from 'redux/modules/app'
@@ -28,7 +28,22 @@ export default class Contact extends Component {
                         status={statuses[SUBMIT_LEAD]}
                         submit={submit}
                     />
-                    <Gallery/>
+                    <Gallery
+                        photos={[
+                            '51QcRqMjy6w',
+                            'v3QeAZjxxtY',
+                            'EeDaZ4So3R0',
+                            'bNGdS51jAoE',
+                            '70RHiKO9kj8',
+                            'C7gNqbkKOtE',
+                            'm5lNqXBBIy0',
+                            'yB-fV93YL54',
+                            '7bVMdNYzH_8',
+                            '_BLx6ystEhk',
+                            'pcbuJpqHUak',
+                            'jjj1rHyYyG0'
+                        ]}
+                    />
                 </Page.Content>
             </Page>
         )
