+++ src/components/Gallery.js
@@ -1,0 +1,25 @@
+import { default as React } from 'react'
+import Gallery from '@leadgrabr/gallery'
+import { Page } from 'components'
+
+const Container = () =>
+    <Page>
+        <Gallery
+            photos={[
+                '51QcRqMjy6w',
+                'v3QeAZjxxtY',
+                'EeDaZ4So3R0',
+                'bNGdS51jAoE',
+                '70RHiKO9kj8',
+                'C7gNqbkKOtE',
+                'm5lNqXBBIy0',
+                'yB-fV93YL54',
+                '7bVMdNYzH_8',
+                '_BLx6ystEhk',
+                'pcbuJpqHUak',
+                'jjj1rHyYyG0'
+            ]}
+        />
+    </Page>
+
+export default Container
