--- src/components/Gallery/Container.js
@@ -1,10 +1,0 @@
-import { default as React } from 'react'
-import Gallery from './Gallery'
-import { Page } from '..'
-
-const Container = () =>
-    <Page>
-        <Gallery/>
-    </Page>
-
-export default Container
