--- src/components/Gallery/index.js
@@ -1,2 +1,0 @@
-import { default as Gallery } from './Gallery'
-export default Gallery
