+++ src/components/index.js
@@ -3,7 +3,7 @@ export { default as App } from './App'
 export { default as BottomBar } from './BottomBar'
 export { default as Contact } from './Contact'
 export { default as DevTools } from './DevTools'
-export { default as Gallery } from './Gallery/Container'
+export { default as Gallery } from './Gallery'
 export { default as GoogleMap } from './GoogleMap'
 export { default as Home } from './Home/Home'
 export { default as Html } from './Html'
